+++ stencil_benchmarks/scripts/sbench_v100_collection.py
@@ -21,7 +21,8 @@ common_kwargs = default_kwargs(backend='cuda',
                                verify=False,
                                dry_runs=1,
                                gpu_timers=True,
-                               alignment=128)
+                               alignment=128,
+                               dtype='float32')
 
 
 @main.command()
